refactor: split TaggedEditOp fixer-glue out of engine::edit_op (#8425)
engine/edit_op.rs mixed two concerns: pure engine primitives (EditOp,
InsertAnchor enums + the apply path) and fixer-pipeline glue (TaggedEditOp +
9 conversion fns from refactor/audit output). The glue pulled
`crate::code_audit::AuditFinding` and `crate::refactor::auto::*` into engine —
the foundation reaching up into feature layers, an inverted dependency arrow.

Split it:
- engine/edit_op.rs keeps only EditOp + InsertAnchor (zero refactor/audit deps).
- apply_edit_ops() now takes &[EditOp] instead of &[TaggedEditOp]; it only ever
  read .op anyway (the primitive/finding/description tags are for reporting,
  not application). Refactor callers pass &t.op.
- TaggedEditOp + all conversions (Insertion/Fix/NewFile/PropagateEdit/
  TransformMatch/FileRename -> TaggedEditOp) move to
  refactor/edit_op_tagged.rs, which is where AuditFinding and RefactorPrimitive
  already live.

Removes the engine::edit_op -> code_audit and engine::edit_op -> refactor
up-edges. Behavior-preserving (moved code + call-site .op mapping). Prerequisite
for both homeboy-engine and homeboy-audit extraction.

Note: a separate engine -> code_audit::conventions::Language edge remains
(edit_op_apply, refactor_primitive) — Language is another mislocated primitive,
tracked as follow-up.

Refs #8425
